\name{plotDiag}
\alias{plotDiag}
\title{Auxiliary function}
\usage{
  plotDiag(nodes, plotnumber)
}
\arguments{
  \item{nodes}{character string, the name of parameters(s)}

  \item{plotnumber}{single numerical value (1 for plotting
  Gelman-Rubin convergence statistic (\code{samplesBgr()}),
  2 for plotting history (\code{plotDensity}) and 3 for
  plotting autocorrelation plots (\code{plotAutoC}))}
}
\description{
  Auxiliary function function for plotting graphical
  convergence diagnostics for Bayes models
}
\keyword{manip}

