\docType{methods}
\name{show-methods}
\alias{show-methods}
\alias{show,bayesmodelClass-method}
\title{Show method for bayesmodelClass}
\arguments{
  \item{object}{a \code{bayesmodelClass} object}
}
\description{
  Show method for \code{\linkS4class{bayesmodelClass}}
}
\examples{
show(new("bayesmodelClass"))
}

